dl-lab-21w-team09-master/human_activity_recognition/requirements.txt
absl==0.0
absl_py==0.12.0
gin==0.1.006
gin_config==0.5.0
matplotlib==3.4.2
numpy==1.21.4
pandas==1.3.4
Pillow==9.0.1
scikit_learn==1.0.2
scipy==1.6.3
seaborn==0.11.2
tensorflow==2.7.0
tensorflow_datasets==4.4.0
tqdm==4.62.3
